improved test coverage, removed vcr, refactored sunlight services
As part of our increased test coverage, we removed VCR. We're testing
APIs that we can't rely on to return the same data in the future, and we
don't want our tests to fail if we delete our VCR cassettes. Since VCR
is an innappropriate tool for what we're testing, we removed it entirely
in favor of using WebMock directly.

We've stopped using environment variables in our tests, as we don't
actually want to hit the *real* API, and if we accidentally did go to
the real internet, we don't want to use real credentials and tweet angry
things at Ted Cruz. I've learned that having application.yml in my
source repository is not a good idea, and will be removing it soon. As a
side effect of our use of real credentials, our tests were failing
because I had a typo in one of the API keys.

The helper method we removed from `ApplicationHelper` was basically just
re-implementing the `to_sentence` method from Active Support, so we're
just using that instead.